Move V2 assessments/evaluation entities into evaluation/ module
