+++ mlflow/evaluation/__init__.py
@@ -1,5 +1,4 @@
-from mlflow.entities.assessment_source import AssessmentSource, AssessmentSourceType
-from mlflow.evaluation.assessment import Assessment
+from mlflow.evaluation.assessment import Assessment, AssessmentSource, AssessmentSourceType
 from mlflow.evaluation.evaluation import Evaluation
 from mlflow.evaluation.fluent import log_evaluations
