+++ mlflow/evaluation/fluent.py
@@ -1,8 +1,7 @@
 import uuid
 from typing import Optional
 
-from mlflow.entities.evaluation import Evaluation as EvaluationEntity
-from mlflow.evaluation.evaluation import Evaluation
+from mlflow.evaluation.evaluation import Evaluation, EvaluationEntity
 from mlflow.evaluation.utils import evaluations_to_dataframes
 from mlflow.tracking.client import MlflowClient
 from mlflow.tracking.fluent import _get_or_start_run
