+++ tests/evaluate/logging/test_assessment.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from mlflow.entities.assessment_source import AssessmentSource
+from mlflow.evaluation.assessment import AssessmentSource
 from mlflow.evaluation import Assessment
 from mlflow.exceptions import MlflowException
