+++ tests/evaluate/logging/test_evaluation.py
@@ -1,8 +1,8 @@
 from unittest.mock import patch
 
 from mlflow.entities import Metric
-from mlflow.entities.assessment_source import AssessmentSource
-from mlflow.entities.evaluation_tag import EvaluationTag
+from mlflow.evaluation.assessment import AssessmentSource
+from mlflow.evaluation.evaluation_tag import EvaluationTag
 from mlflow.evaluation import Assessment, Evaluation
